requirements.txt
# Web Framework | Web 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Telegram Client | Telegram 客戶端
telethon>=1.34.0

# Template Engine | 模板引擎
jinja2>=3.1.3

# Environment Variables | 環境變數
python-dotenv>=1.0.0

# YAML Support | YAML 支援
pyyaml>=6.0.1

# Security | 安全性
cryptography>=42.0.0
